Miss-Click encountered, object Spawning increased
I have added a code that whenever user missclick on the wall means, while clicking the object he miss it and click the wall, 5 points are reduced. and also increased the object spawning in different difficulties, like in easy there is 10% probability that 2 objects will be spawned instead of one at a time, in medium, there is 25% probability, and in hard 50% probability.